HIVE-24599: Extract StringTrimScalarColBase for SCALAR+COLUMN trim
